refactor: MapContainer에서 MapView와 useMapData 훅으로 분리
